Refactored services file + Added environments + Minor fixes
