+++ src/app/chat/chat-assistant/chat-assistant.component.ts
@@ -1,6 +1,6 @@
 import { Component, ElementRef, inject, ViewChild } from '@angular/core';
 import { ChatMessage } from '../../models/chat-assistant.model';
-import { ChatAssistantService } from '../../services/chat-assistant.service';
+import { ChatAssistantService } from '../../services/chat-assistant.service/chat-assistant.service';
 import { MarkdownModule } from 'ngx-markdown';
 @Component({
   selector: 'app-chat-assistant',
