+++ src/app/services/chat-assistant.service/chat-assistant.service.ts
@@ -1,13 +1,14 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { map, Observable } from 'rxjs';
+import { environment } from '../../../environments/environment';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ChatAssistantService {
 
-  private apiUrl = 'http://127.0.0.1:5000'
+  private apiUrl = environment.apiUrl;
 
   constructor(private http: HttpClient) {}
