+++ src/app/services/parcel-finder.service/parcel-finder.service.ts
@@ -1,14 +1,15 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { map, Observable } from 'rxjs';
-import { cropClassification } from '../models/parcel-finder.models';
+import { cropClassification } from '../../models/parcel-finder.models';
+import { environment } from '../../../environments/environment';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ParcelFinderService {
   
-  private apiUrl = 'http://127.0.0.1:5000'
+  private apiUrl = environment.apiUrl;
 
   constructor(private http: HttpClient) {}
