+++ src/environments/environment.development.ts
@@ -1,0 +1,1 @@
+export const environment = {};
