+++ src/environments/environment.ts
@@ -1,0 +1,5 @@
+export const environment = {
+  production: true,
+  apiUrl: 'http://127.0.0.1:5000', // Your production API URL
+  apiPort: 443 // Standard HTTPS port for production
+};
